requirements.txt
automata-lib
pandas
graphviz
colormath
jupyterlab
visual-automata